+++ src/main/java/com/dfsek/terra/biome/UserDefinedGrid.java
@@ -15,7 +15,7 @@ public class UserDefinedGrid extends BiomeGrid {
     private final ImageLoader.Channel channelX;
     private final ImageLoader.Channel channelZ;
 
-    public UserDefinedGrid(World w, float freq1, float freq2, UserDefinedBiome[][] b, WorldConfig c) {
+    public UserDefinedGrid(World w, double freq1, double freq2, UserDefinedBiome[][] b, WorldConfig c) {
         super(w, freq1, freq2, b.length, b[0].length);
         super.setGrid(b);
         imageLoader = c.imageLoader;
