+++ src/main/java/com/dfsek/terra/image/WorldImageGenerator.java
@@ -28,7 +28,7 @@ public class WorldImageGenerator {
             int imX = 0;
             for(int x = centerX - (draw.getWidth() / 2); x < centerX + (draw.getWidth() / 2); x++) {
                 int zone = NormalizationUtil.normalize(tw.getZone().getRawNoise(x, y), 256, 4);
-                float[] noise = tb.getGrid(x, y).getRawNoise(x, y);
+                double[] noise = tb.getGrid(x, y).getRawNoise(x, y);
                 Color c = new Color(NormalizationUtil.normalize(noise[0], 256, 4), NormalizationUtil.normalize(noise[1], 256, 4), zone);
                 draw.setRGB(imX, imY, c.getRGB());
                 imX++;
